bigslice/exec: wrap task errors with more context
Summary: Wrap task errors with more context to make it easier to diagnose what is happening when they are returned from evaluations. We now provide the task name and invocation to give clues as to what might be going wrong with evaluation, with less need to sift through mountains of logs.

Test Plan: Unit test

Reviewers: pgopal, marius

Reviewed By: marius

Differential Revision: https://phabricator.grailbio.com/D37259

fbshipit-source-id: 8df99db