bigslice/exec: stop indirect machine failures from causing probation
Summary:
Stop indirect machine failures from causing probation. For example, suppose we run task `T` on machine `M`, and `T` must read data from machine `N`. If `N` is unhealthy (e.g. unavailable, unresponsive, etc.), we do not want failures reaching `N` to signal that `M` is unhealthy.

We previously accomplished this by having special cases around `errors.Net` and `errors.Unavailable` kinds. However, this was imprecise.

In this revision, we add a new error kind to allows clients of `bigmachine.Call` to differentiate between errors in RPC execution machinery and application errors from the RPC itself. `bigslice` uses this distinction to only put a machine on probation if it directly observes errors in RPC machinery for that machine.

This simplifies the error taxonomy for task execution considerably.  An error in task execution can be `errors.Fatal`, in which case we do not retry. If it is not `errors.Fatal`, we will retry execution of that task.  If the non-fatal error comes from application code (`errors.RPC`), we'll consider the machine healthy, otherwise we'll consider the machine unhealthy.

Test Plan: Unit test and manual test

Reviewers: pgopal, marius

Reviewed By: marius

Subscribers: joshnewman, psingh, smahadevan

Differential Revision: https://phabricator.grailbio.com/D37368

fbshipit-source-id: ab55f21