bigslice/exec: retry calls to Worker.Stat
Summary: Retry calls to `Worker.Stat`, like we retry calls for everything else. As is, this will fall back to the task execution machinery to retry the lost task, which is okay except that with D37256, we'll give up on the task too soon on known-temporary errors.

Test Plan: Unit test

Reviewers: pgopal, marius

Reviewed By: marius

Differential Revision: https://phabricator.grailbio.com/D37654

fbshipit-source-id: e2118c2