bigslice/exec: run exclusive slices on existing clusters
Summary:
Run exclusive slices on existing clusters, instead of on separate reserved clusters. Do this by allowing tasks to require an arbitrary number of procs on machines -- if the `bigslice.Exclusive` is set, the task is set to require the maximum available number of procs on a single machine, so it will be scheduled exclusively.

Change scheduling to be first fit decreasing to accommodate the heterogeneous proc requirements.

This opens the door to tasks having arbitrary proc requirements, but that is not yet needed/implemented.

Test Plan: Update and run unit tests. Run classifier evaluation manual test.

Reviewers: pgopal, marius

Reviewed By: marius

Subscribers: jschellenberger

Maniphest Tasks: T20118

Differential Revision: https://phabricator.grailbio.com/D38141

fbshipit-source-id: a70ae96