slicefunc: package to abstract user-provided funcs
Package slicefunc abstracts user-provided funcs. We replace
occurences of user-provided functions with the new type
slicefunc.Func. slicefunc.Func manages whether to plumb a context
arguments to the function. (typecheck.Func is amended accordingly.)

Slicefunc can be used in the future to provide additional
functionality (e.g., allowing a user to return an error explicitly),
and also gives us an abstraction through which we may perform future
optimizations.

This is the first step towards #18.