support for user-defined metrics (#24)
* slicefunc: package to abstract user-provided funcs

Package slicefunc abstracts user-provided funcs. We replace
occurences of user-provided functions with the new type
slicefunc.Func. slicefunc.Func manages whether to plumb a context
arguments to the function. (typecheck.Func is amended accordingly.)

Slicefunc can be used in the future to provide additional
functionality (e.g., allowing a user to return an error explicitly),
and also gives us an abstraction through which we may perform future
optimizations.

This is the first step towards #18.

* slicefunc: package to abstract user-provided funcs

Package slicefunc abstracts user-provided funcs. We replace
occurences of user-provided functions with the new type
slicefunc.Func. slicefunc.Func manages whether to plumb a context
arguments to the function. (typecheck.Func is amended accordingly.)

Slicefunc can be used in the future to provide additional
functionality (e.g., allowing a user to return an error explicitly),
and also gives us an abstraction through which we may perform future
optimizations.

This is the first step towards #18.

* metrics: wip

* Update exec/task.go

Co-Authored-By: jcharum <jcharumilind@grailbio.com>

* Update slice.go

Co-Authored-By: jcharum <jcharumilind@grailbio.com>

* Update slicefunc/func.go

Co-Authored-By: jcharum <jcharumilind@grailbio.com>

* metrics: flesh out package

* metrics: a few more things

* integrate metric scopes in evaluators

* Update go.sum

* remove commented out code

* merge nit

* make the atomic code a lot clearer

* Apply suggestions from code review

Co-Authored-By: jcharum <jcharumilind@grailbio.com>

* clarify, simplify

* fix races

* capitalization

* make gob test stricter

* properly handle noninitialized metrics while decoding

* add documentation for (*Scope).store

* empty commit to trigger build

* use nil instead of uintptr(0) to make linters happy