Make compilation stable w.r.t. cache state (#40)
Capture the (cache) environment used for compilation on the driver and propagate it to workers to make compilation stable with respect to cache state, i.e. same task graph is compiled even if workers see different cache state. This should fix panics that we see cause by task X not found errors, in which unstable compilation leads the driver to try to run tasks on workers that have not compiled those tasks.

This includes some tweaks to support unit testing.

I also manually tested this with a test program that exercises this panic by causing machine death once the cache is populated.