Add ReadCache slice operation (#46)
Add a `ReadCache` slice operation that reads from an existing cache but does not write any cache itself. This may be useful if you want to reuse a cache from a previous computation and fail if it does not exist.

This change also removes the errors returned by the `Cache` and `CachePartial` operations, as they are never non-`nil`. They previously were used to signal an incomplete cache, but this is now handled by the `IsCached` method. This simplifies callers, especially since these are virtually always called within `bigslice.Func` definitions which do not an easy way to propagate errors.