Add useful durations to trace event arguments (#58)
Add useful durations to trace event arguments, specifically durations for read and write overhead. Read duration captures the time spent reading from dependencies, which includes transfer and decoding time. Write duration captures the time spent writing results to storage and for combiner task, combiner commit time.

Note this does not capture machine combiner commit time, as this does not happen within the context of individual tasks.

As part of this change, we introduce `sliceio.Writer`, which is the dual of `sliceio.Reader`.