Remove unhelpful log message (#62)
Remove unhelpful log message written when we see an "invocation" event with an invocation index that we've already seen. We actually expect the same invocation index event for each machine. We just assume that they are all the same, and use the information from the first one. The log is unhelpful.