Free ephemeral resources used for scanning (#70)
When we scan results, we get a slice of `*openerAtReader`s, one for each result task. We read from each reader sequentially. When we are done with a given reader, we retain some of the resources used when reading from it, most notably `gob` decode buffers. As we scan, we accumulate these defunct buffers, and our memory footprint grows.

This happens for two reasons:

1. We pop off the slice of readers to iterate, i.e. `q = q[1:]`. However, we do not clear the backing array reference to the reader.
2. When we close the reader, we don't clear the `sliceioReader`, which in turn holds the `gob` decoder.

Fixing either would eliminate the specific scan leak. Fix both, as I think it's the correct behavior.