Reuse Func argument gob-encoding (#87)
Reuse `Func` argument gob-encoding for `Worker.Compile` calls. Arguments may be large. Gob-encoding uses an internal buffer for decoding that's at least the size of the value being encoded. We may call `Worker.Compile` for many workers concurrently. This can use excessive memory, e.g. 500MB * 50 workers. We have seen this lead to OOMs with real workloads.

By reusing, we hold on to the gob-encoding and reuse it for all calls for that invocation. We end up holding on to this buffer, but we avoid the spikes of concurrent encoding.

This PR also refactors the compilation environment code to make the coupling to invocation more clear, as we rely on the environment being write-once for a given invocation.