Make new reader for invocation arg for each retry (#94)
Switch to use the `func() io.Reader` argument type for the RPC of `Worker.Compile`. By doing this, each attempt to make the RPC (retried on temporary failures) will transmit the entire encoded invocation. Before this commit, retry attempts to make the RPC would likely fail because the reader had already been (partially) read by the previous invocation. Instead, the `func` we pass in will be called for each retry attempt, and it will construct a reader that reads out the entire invocation.

Upgrade `bigmachine` to `v.0.5.8`, as that version supports `func() io.Reader` RPC arguments.